requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.95.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.2
pyasn1==0.4.8
pydantic==1.10.7
pypika-tortoise==0.1.6
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.21.1
